unknown: sh[ia]me?le[es]s word
There is not enough context to figure out what word is intended.

It could be: shimless, shameless, or something else...

As there are two words which seem to be used in a similar context, but
with slightly different spellings, I assert they are the same word.

But I can not offer a correction.